requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
emoji==1.7.0
feedparser==6.0.11
future==0.18.3
idna==3.6
multicolorcaptcha==1.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
python-telegram-bot==11.1.0
requests==2.31.0
sgmllib3k==1.0.0
SQLAlchemy==1.3.23
urllib3==2.2.0
